refactor: refactor stonedb aggregate function
