+++ src/sqlancer/stonedb/oracle/StoneDBAggregateOracle.java
@@ -19,9 +19,9 @@ import sqlancer.stonedb.StoneDBProvider.StoneDBGlobalState;
 import sqlancer.stonedb.StoneDBSchema.StoneDBCompositeDataType;
 import sqlancer.stonedb.StoneDBSchema.StoneDBDataType;
 import sqlancer.stonedb.StoneDBToStringVisitor;
+import sqlancer.stonedb.ast.StoneDBAggregate.StoneDBAggregateFunction;
 import sqlancer.stonedb.ast.StoneDBExpression;
 import sqlancer.stonedb.ast.StoneDBSelect;
-import sqlancer.stonedb.gen.StoneDBExpressionGenerator.StoneDBAggregateFunction;
 import sqlancer.stonedb.gen.StoneDBExpressionGenerator.StoneDBCastOperation;
 import sqlancer.stonedb.gen.StoneDBExpressionGenerator.StoneDBUnaryPostfixOperator;
 import sqlancer.stonedb.gen.StoneDBExpressionGenerator.StoneDBUnaryPrefixOperator;
